+++ common/src/main/java/dev/ftb/mods/ftbquests/net/ClaimAllRewardsMessage.java
@@ -3,17 +3,21 @@ package dev.ftb.mods.ftbquests.net;
 import dev.architectury.networking.NetworkManager;
 import dev.ftb.mods.ftbquests.api.FTBQuestsAPI;
 import dev.ftb.mods.ftbquests.quest.ServerQuestFile;
+import dev.ftb.mods.ftbquests.quest.reward.Reward;
+import dev.ftb.mods.ftbquests.quest.reward.RewardTypes;
 import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import net.minecraft.server.level.ServerPlayer;
 
-public class ClaimAllRewardsMessage implements CustomPacketPayload {
+public record ClaimAllRewardsMessage(boolean xpOnly) implements CustomPacketPayload {
 	public static final Type<ClaimAllRewardsMessage> TYPE = new Type<>(FTBQuestsAPI.rl("claim_all_rewards_message"));
 
-	public static final ClaimAllRewardsMessage INSTANCE = new ClaimAllRewardsMessage();
-
-	public static final StreamCodec<FriendlyByteBuf, ClaimAllRewardsMessage> STREAM_CODEC = StreamCodec.unit(INSTANCE);
+	public static final StreamCodec<FriendlyByteBuf, ClaimAllRewardsMessage> STREAM_CODEC = StreamCodec.composite(
+			ByteBufCodecs.BOOL, ClaimAllRewardsMessage::xpOnly,
+			ClaimAllRewardsMessage::new
+	);
 
 	@Override
 	public Type<ClaimAllRewardsMessage> type() {
@@ -23,15 +27,19 @@ public class ClaimAllRewardsMessage implements CustomPacketPayload {
 	public static void handle(ClaimAllRewardsMessage message, NetworkManager.PacketContext context) {
 		context.queue(() -> {
 			ServerPlayer player = (ServerPlayer) context.getPlayer();
-			ServerQuestFile.INSTANCE.getTeamData(player).ifPresent(data -> {
-				data.getFile().forAllQuests(quest -> {
-					if (data.isCompleted(quest)) {
-						quest.getRewards().stream()
-								.filter(reward -> !reward.getExcludeFromClaimAll())
-								.forEach(reward -> data.claimReward(player, reward, true));
-					}
-				});
-			});
+			ServerQuestFile.INSTANCE.getTeamData(player)
+					.ifPresent(data -> data.getFile().forAllQuests(quest -> {
+						if (data.isCompleted(quest)) {
+							quest.getRewards().stream()
+									.filter(message::shouldIncludeReward)
+									.forEach(reward -> data.claimReward(player, reward, true));
+						}
+					}));
 		});
+	}
+
+	private boolean shouldIncludeReward(Reward reward) {
+		return !reward.getExcludeFromClaimAll()
+				&& (!xpOnly || reward.getType() == RewardTypes.XP || reward.getType() == RewardTypes.XP_LEVELS);
 	}
 }